+++ physics_update.c
@@ -7,7 +7,7 @@
 
 void physics_update( game_data * gd, float dt){
   game_entities * entities = &gd->entities;
-  for(size_t i = 0; i < entities->cnt; i++){
+  for(int i = 0; i < entities->cnt; i++){
     entities->x[i] += entities->dx[i] * dt;
     entities->y[i] += entities->dy[i] * dt;
   }
